+ed rock, paper, scissors
The rps.py program for the rock paper scissors game has a bug when the user selects paper, the logic in the relevant elif branch is incorrect we'll be fixing this bug.
As part of another issue you'll need to give the user the ability to repeatedly play the game until they decide to quit.  This can be done with a loop, another call to input() and a break if the user wants to quit.